+++ src/Phan/Language/Type.php
@@ -809,13 +809,15 @@ class Type
             }
             return $reflection_type_string;
         }
-        // Unreachable in php 7.1+? Also, ReflectionType::__toString() is deprecated in php 7.4
+        // Unreachable in php 7.1-7.4, but reachable and revertsed deprecation in php 8.0+?
         return (string)$reflection_type;
     }
 
     /**
      * Creates a type for the ReflectionType of a parameter, return value, etc.
      * @phan-side-effect-free
+     * @deprecated - use UnionType::fromReflectionType to prepare for php 8
+     * @suppress PhanUnreferencedPublicMethod
      */
     public static function fromReflectionType(
         \ReflectionType $reflection_type